pyproject.toml
[project]
name = "simple-online-taxi"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "zparto82", email = "zparto82@gmail.com" }]
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "alembic>=1.15.2",
    "pydantic[email]>=2.11.4",
    "aiosqlite>=0.21.0",
    "sqlalchemy[asyncio]>=2.0.40",
    "python-jose[cryptography]>=3.4.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.9.1",
    "protobuf>=6.30.2",
    "valkey>=6.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
virtual = true
managed = true
dev-dependencies = ["ruff>=0.11.7"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/simple_online_taxi"]

[tool.rye.scripts]
dev = "uvicorn src.simple_online_taxi.main:app --reload"
test = "python tests/unit_test.py"
prod = "uvicorn src.simple_online_taxi.main:app --workers 2"
migrate = "alembic upgrade head"
makemigrations = { cmd = "python -c \"import datetime; desc=input('Migration description: '); msg=desc or f'Autogen {datetime.datetime.now().strftime('%Y-%m-%d %H:%M')}'; import subprocess; subprocess.run(['alembic', 'revision', '--autogenerate', '-m', msg])\"", shell = true }
